desktop-app/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
openpyxl==3.1.2
gspread==5.12.0
google-auth==2.23.4
pydantic==2.5.0
python-multipart==0.0.6
